data: cast windowDays to int in active-roles date window
Postgres couldn't resolve 'current_date + $1' when pg bound $1 as
type 'unknown' — error: 'operator is not unique: date + unknown'.
Adding an explicit ($1)::int cast picks the date+int operator
unambiguously.

Reproduced on production with Section 8 (Forward Signal) triggering
an SSR 500 on every page render; digest 2136610193 in the HTML.

+++ src/lib/data.ts
@@ -253,7 +253,10 @@ export async function getActiveRoles(filter?: {
   }
   if (filter?.windowDays) {
     conditions.push(
-      sql`${activeRoles.closingDate} is not null and ${activeRoles.closingDate} <= current_date + ${filter.windowDays}`,
+      // pg binds windowDays as 'unknown', which makes `date + unknown`
+      // ambiguous to Postgres ("operator is not unique"). Cast explicitly to
+      // int so the date + int operator is selected.
+      sql`${activeRoles.closingDate} is not null and ${activeRoles.closingDate} <= current_date + (${filter.windowDays})::int`,
     );
   }
   const whereClause =
